Warning fixes.

+++ modules/features2d/src/kaze.cpp
@@ -75,7 +75,7 @@ namespace cv
         void setUpright(bool upright_) CV_OVERRIDE { upright = upright_; }
         bool getUpright() const CV_OVERRIDE { return upright; }
 
-        void setThreshold(double threshold_) CV_OVERRIDE { threshold = threshold_; }
+        void setThreshold(double threshold_) CV_OVERRIDE { threshold = (float)threshold_; }
         double getThreshold() const CV_OVERRIDE { return threshold; }
 
         void setNOctaves(int octaves_) CV_OVERRIDE { octaves = octaves_; }
